requirements.txt
streamlit
sqlalchemy
bcrypt
pandas
pdfplumber
rank-bm25
python-docx
openai
tqdm